[feat/mingi] Create Yaml files and environment
