+++ src/main/resources/application-dev.yaml
@@ -1,0 +1,3 @@
+spring:
+  application:
+    name: discodeit
