+++ src/main/resources/application.yaml
@@ -1,0 +1,9 @@
+spring:
+  application:
+    name: discodeit
+
+  datasource:
+    url: jdbc:mysql://localhost:3306/discodeitDB
+    username: root
+    password: 1234
+    driver-class-name: com.mysql.cj.jdbc.Driver
