+++ src/test/java/com/sprint/mission/discodeit/DiscodeitApplicationTests.java
@@ -1,0 +1,13 @@
+package com.sprint.mission.discodeit;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class DiscodeitApplicationTests {
+
+    @Test
+    void contextLoads() {
+
+    }
+}
